Add prop for columns
Currently the number of columns is hard-coded, so that - when people use different slot content for this component - the number of columns does not match up. With this PR, people can simply pass in a prop from the parent component (or a wrapper component `src/components/organisms/SwFooterNavigation.vue`) to change the number.